galileo-a2a/pyproject.toml
[project]
name = "galileo-a2a"
version = "1.0.0b1"
description = "Galileo observability auto-instrumentation for A2A (Agent-to-Agent) protocol"
readme = "README.md"
requires-python = ">=3.10,<3.15"
license = { text = "Apache-2.0" }
authors = [{ name = "Galileo Technologies Inc.", email = "team@galileo.ai" }]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "galileo>=1.47.0,<2.0.0",
    "a2a-sdk>=0.3.0,<1.0.0",
    "opentelemetry-api>=1.38.0",
    "opentelemetry-sdk>=1.38.0",
]

[project.urls]
Repository = "https://github.com/rungalileo/galileo-python"

# UV-specific configuration (path for dev, ignored when installed from PyPI)
[tool.uv]
sources = { galileo = { path = "../", editable = true } }

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.0.0",
    "pytest-xdist>=3.7.0",
    "pytest-socket>=0.7",
    "ruff>=0.12.3",
    "mypy>=1.16.0",
    "coverage>=7.9.2",
    "galileo-core[testing]>=3.82.0",
    "uvicorn>=0.42.0",
    "starlette>=1.0.0",
    "httpx>=0.28.1",
    "opentelemetry-instrumentation>=0.61b0",
    "galileo[otel]>=1.50.1",
    "sse-starlette>=3.3.4",
    "openai>=2.30.0",
    "opentelemetry-instrumentation-openai>=0.57.0",
    "opentelemetry-instrumentation-langchain>=0.57.0",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "python-dotenv>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/galileo_a2a"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
env = [
    "GALILEO_CONSOLE_URL=http://localtest:8088",
    "GALILEO_API_KEY=api-1234567890",
    "GALILEO_PROJECT=test-project",
    "GALILEO_LOG_STREAM=test-log-stream",
]
addopts = [
    "-v",
    "--strict-markers",
    "--disable-warnings",
    "--showlocals",
]

[tool.ruff]
line-length = 120
fix = true
src = ["src"]
target-version = "py310"

[tool.ruff.lint]
select = ["F", "E4", "E7", "E9", "W6", "I", "UP", "ASYNC", "B", "C4", "ARG", "PT", "RET", "SIM"]
ignore = [
    "S101",
    "ARG001",
    "ARG002",
    "D",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ANN",
    "PLR",
    "S",
    "ARG",
    "D",
]

[tool.ruff.lint.isort]
known-first-party = ["galileo_a2a", "galileo"]

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
follow_imports = "skip"

[tool.coverage.run]
source = ["src/galileo_a2a"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

# =============================================================================
# Release Configuration
# =============================================================================

[tool.semantic_release]
version_variables = ["src/galileo_a2a/__init__.py:__version__"]
version_toml = ["pyproject.toml:project.version"]
tag_format = "galileo-a2a-v{version}"
version_source = "tag"
commit_message = "chore(release): galileo-a2a v{version}\n\nAutomatically generated by python-semantic-release"

[tool.semantic_release.commit_parser_options]
patch_tags = ["fix", "perf", "chore", "docs", "style", "refactor"]

[tool.semantic_release.remote]
type = "github"

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
